put_hexa has long as d_type, it is changed to int for correct implementation

+++ ft_printf.h
@@ -27,7 +27,7 @@ int	print_args(char *specifiers, va_list args);
 int	char_group(char specifier, va_list args);
 int	int_group(char specifier, va_list args);
 int	put_ints_in_hexa(char specifier, va_list args);
-char	*decimal_to_hexadecimal(long integer, char letter_size);
+char	*decimal_to_hexadecimal(int integer, char letter_size);
 int	puthexa(int num, char size);
 
 #endif
